requirements.txt
mcp[cli]
gooddata-sdk
python-dotenv
